refactor: merge duplicate implementation of boolean matcher
