+++ crates/rspack_plugin_runtime/src/lib.rs
@@ -21,9 +21,7 @@ pub use module_chunk_loading::ModuleChunkLoadingPlugin;
 mod import_scripts_chunk_loading;
 pub use import_scripts_chunk_loading::ImportScriptsChunkLoadingPlugin;
 mod runtime_module;
-pub use runtime_module::{
-  chunk_has_css, is_enabled_for_chunk, render_condition_map, stringify_chunks, BooleanMatcher,
-};
+pub use runtime_module::{chunk_has_css, is_enabled_for_chunk, stringify_chunks};
 mod startup_chunk_dependencies;
 pub use startup_chunk_dependencies::StartupChunkDependenciesPlugin;
 mod chunk_prefetch_preload;
