+++ crates/rspack_plugin_runtime/src/runtime_module/module_chunk_loading.rs
@@ -1,16 +1,12 @@
 use rspack_core::{
-  impl_runtime_module,
+  compile_boolean_matcher, impl_runtime_module,
   rspack_sources::{BoxSource, ConcatSource, RawSource, SourceExt},
-  Chunk, ChunkUkey, Compilation, RuntimeGlobals, RuntimeModule, RuntimeModuleStage,
+  BooleanMatcher, Chunk, ChunkUkey, Compilation, RuntimeGlobals, RuntimeModule, RuntimeModuleStage,
 };
 use rspack_identifier::Identifier;
 use rspack_util::source_map::SourceMapKind;
 
-use super::{
-  render_condition_map,
-  utils::{chunk_has_js, get_output_dir},
-  BooleanMatcher,
-};
+use super::utils::{chunk_has_js, get_output_dir};
 use crate::{
   get_chunk_runtime_requirements,
   runtime_module::utils::{get_initial_chunk_ids, stringify_chunks},
@@ -77,7 +73,7 @@ impl RuntimeModule for ModuleChunkLoadingRuntimeModule {
       compilation
         .chunk_graph
         .get_chunk_condition_map(&chunk.ukey, compilation, chunk_has_js);
-    let has_js_matcher = render_condition_map(&condition_map, "chunkId");
+    let has_js_matcher = compile_boolean_matcher(&condition_map);
     let initial_chunks = get_initial_chunk_ids(self.chunk, compilation, chunk_has_js);
 
     let root_output_dir = get_output_dir(chunk, compilation, true);
@@ -122,7 +118,7 @@ impl RuntimeModule for ModuleChunkLoadingRuntimeModule {
         "installedChunks[chunkId] = 0;".to_string()
       } else {
         include_str!("runtime/module_chunk_loading_with_loading.js")
-          .replace("$JS_MATCHER$", has_js_matcher.to_string().as_str())
+          .replace("$JS_MATCHER$", &has_js_matcher.render("chunkId"))
           .replace(
             "$IMPORT_FUNCTION_NAME$",
             &compilation.options.output.import_function_name,
